handle stdin EOF
Reading from a input at end-of-file will write 0 to the current cell.
There is no agreement in what the behaviour should be (keep the cell
value, like the original compiler, write -1, like original interpreter,
or write 0, like many implementations).

Choose to write 0 because the cat.bf program that I was testing with
relies on that.