Add some micro optimizations
Add a bunch of `if program_counter > memory.len() {
unreachable_unchecked() }` to remove bound checks.

Also use a power of 2 length for the memory to allow optimizing the wrap
around logic. (didn't look at the assembly but it worked).

Everything together has decreased runtime from 13.5 to 7.1 seconds on
mandelbrot.bf.